src/requirements.txt
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
Django==2.2.5
djangorestframework==3.10.3
djongo==1.3.6
drf-firebase-auth==1.0.0
firebase-admin==4.5.3
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.39.0
httplib2==0.19.1
idna==2.10
msgpack==1.0.2
numpy==1.18.3
packaging==21.0
pandas==1.0.3
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.23.0
rsa==4.7.2
six==1.16.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.25.11
wincertstore==0.2
redis
celery
djongo-celery-results
uwsgi